merge the Universe instances for [a] and InfiniteList a
The InfiniteList a instance has appropriate laziness properties, and
relaxes the "Finite a" constraint to be merely "Universe a", so it seems
more desirable to me. It has a slightly different ordering than the old
instance, so it may have different memory characteristics; however, both
implementations seem like they're going to be absolute hell in that
department. So instead of supporting both instances indefinitely, let's
support just the one that seems better for now until somebody complains.