Improve Rational instance
Use a stream
This feels cleaner to me, but it's not a big deal either way.

Only use the left side of the tree

It turns out that we can save a lot of time and improve sharing by
building only the traversal of the left side of the Calkin-Wilf tree. We
can then generate four rational numbers from each node, rather than just
two. This means we perform fewer sums, so less allocation, and we also
get more from each pricy pointer lookup.

Note: this commit changes the order of elements in
`universe @Rational`, but hopefully no one is relying on it
staying exactly the same forever.

Take the reciprocals manually to avoid unnecessary sign tests.

Add test
Make the test work

Just throwing `Arbitrary` rationals leads to tests that quickly
run out of memory. Limit the ranges to avoid this.

Add instances for Natural and Ratio Natural

Now that `Natural` is in `base`, I think both `Natural` and
`Ratio Natural` deserve `Universe` instances. Add them.