Only use the left side of the tree
It turns out that we can save a lot of time and improve sharing by
building only the traversal of the left side of the Calkin-Wilf tree. We
can then generate four rational numbers from each node, rather than just
two. This means we perform fewer sums, so less allocation, and we also
get more from each pricy pointer lookup.

Note: this commit changes the order of elements in
`universe @Rational`, but hopefully no one is relying on it
staying exactly the same forever.